backend/requirements.txt
Django==3.2.15
django-debug-toolbar==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0

